event_registration/requirements.txt
Django==5.0
django-import-export==3.3.5
